requirements.txt
aiohttp==3.9.3
aioselenium==0.0.1
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors==0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-jazzmin==3.0.0
django-tables2==2.7.0
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
python-decouple
psycopg2-binary
gunicorn
drf-spectacular==0.27.1
setuptools
django-ckeditor
django-modeltranslation==0.18.11
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
